Postgress_7/requirements.txt
jwt
fastapi
pydantic
datetime
typing
fastapi
tortoise-orm
uvicorn
asyncpg
